# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SpendShield API — subscription analyzer and savings tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List all detected subscriptions for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: addSubscription
      tags: [subscriptions]
      summary: Manually add a subscription the detector missed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSubscriptionInput"
      responses:
        "201":
          description: created subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get a single subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update subscription status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: Not found
  /subscriptions/{id}/message:
    get:
      operationId: getSubscriptionMessage
      tags: [subscriptions]
      summary: Get the saved negotiation/cancellation message for a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Negotiation message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NegotiationMessage"
        "404":
          description: No message generated yet
    post:
      operationId: generateSubscriptionMessage
      tags: [subscriptions]
      summary: Generate a negotiation or cancellation message for a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"
      responses:
        "200":
          description: Generated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NegotiationMessage"
        "404":
          description: Subscription not found
  /subscriptions/bundle-suggestions:
    get:
      operationId: getBundleSuggestions
      tags: [subscriptions]
      summary: Bundle Optimizer — flags categories with 2+ active subscriptions worth consolidating
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bundle suggestions grouped by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BundleSuggestion"
...
  /subscriptions/deal-watch:
    get:
      operationId: getDealWatch
      tags: [subscriptions]
      summary: Deal Watch — curated deals/offers matching the user's active subscriptions
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matched deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DealMatch"
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, sessionId, name, amount, frequency, category, status, keepCount, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        name:
          type: string
        amount:
          type: number
        frequency:
          type: string
          enum: [monthly, annual, weekly, unknown]
        category:
          type: string
          enum: [streaming, fitness, software, food, finance, utility, gaming, education, other]
        status:
          type: string
          enum: [active, flagged, cancelled, reviewing]
        flagReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        previousAmount:
          type: ["number", "null"]
          description: Set automatically when a price hike is detected on update
        keepCount:
          type: integer
          description: Times user dismissed a flag and kept this subscription active
        createdAt:
          type: string
    SubscriptionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, flagged, cancelled, reviewing]
        notes:
          type: string
        amount:
          type: number
          description: Updating this triggers Price-Hike Detector if higher than the current amount
    NegotiationMessage:
      type: object
      required: [id, subscriptionId, messageType, message, createdAt]
      properties:
        id:
          type: integer
        subscriptionId:
          type: integer
        messageType:
          type: string
          enum: [cancel, negotiate, downgrade, refund]
        message:
          type: string
        serviceName:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
          enum: [worked, partial, ignored, declined, null]
        createdAt:
          type: string
    MessageRequest:
      type: object
      required: [messageType]
      properties:
        messageType:
          type: string
          enum: [cancel, negotiate, downgrade, refund]
...
    BundleSuggestion:
      type: object
      required: [category, subscriptions, combinedMonthlyCost, message]
      properties:
        category:
          type: string
        subscriptions:
          type: array
          items:
            type: string
        combinedMonthlyCost:
          type: number
        message:
          type: string
...
    DealMatch:
      type: object
      required: [subscriptionName, title, description]
      properties:
        subscriptionName:
          type: string
        title:
          type: string
        description:
          type: string
...
    AddSubscriptionInput:
      type: object
      required: [sessionId, name, amount, frequency, category]
      properties:
        sessionId:
          type: string
        name:
          type: string
        amount:
          type: number
        frequency:
          type: string
          enum: [monthly, annual, weekly, unknown]
        category:
          type: string
          enum: [streaming, fitness, software, food, finance, utility, gaming, education, other]
...